Cleaned up v1 for app store pages. (#5204)
