+++ packages/app-store/applecalendar/pages/setup/index.tsx
@@ -4,9 +4,7 @@ import { useForm } from "react-hook-form";
 import { Toaster } from "react-hot-toast";
 
 import { useLocale } from "@calcom/lib/hooks/useLocale";
-import { Alert } from "@calcom/ui/Alert";
-import Button from "@calcom/ui/Button";
-import { Form, TextField } from "@calcom/ui/form/fields";
+import { Alert, Button, Form, TextField } from "@calcom/ui/v2";
 
 export default function AppleCalendarSetup() {
   const { t } = useLocale();
