--- packages/app-store/zapier/pages/v2/setup/_getStaticProps.tsx
@@ -1,20 +1,0 @@
-import { GetStaticPropsContext } from "next";
-
-import getAppKeysFromSlug from "../../../../_utils/getAppKeysFromSlug";
-
-export interface IZapierSetupProps {
-  inviteLink: string;
-}
-
-export const getStaticProps = async (ctx: GetStaticPropsContext) => {
-  if (typeof ctx.params?.slug !== "string") return { notFound: true } as const;
-  let inviteLink = "";
-  const appKeys = await getAppKeysFromSlug("zapier");
-  if (typeof appKeys.invite_link === "string") inviteLink = appKeys.invite_link;
-
-  return {
-    props: {
-      inviteLink,
-    },
-  };
-};
